Feat - Implementing "GORM"
- Implemeting "GORM" ORM in to the project

+++ utils/jwt.go
@@ -9,7 +9,7 @@ import (
 
 const secretKey = "supersecret"
 
-func GenerateToken(email string, userId int64) (string, error) {
+func GenerateToken(email string, userId uint) (string, error) {
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
 		"email":  email,
 		"userId": userId,
@@ -19,7 +19,7 @@ func GenerateToken(email string, userId int64) (string, error) {
 	return token.SignedString([]byte(secretKey))
 }
 
-func VerifyToken(token string) (int64, error) {
+func VerifyToken(token string) (uint, error) {
 
 	parsedToken, err := jwt.Parse(token, func(token *jwt.Token) (interface{}, error) {
 		_, ok := token.Method.(*jwt.SigningMethodHMAC)
@@ -45,7 +45,7 @@ func VerifyToken(token string) (int64, error) {
 	}
 
 	// email := claims["email"].(string)
-	userId := int64(claims["userId"].(float64))
+	userId := uint(claims["userId"].(float64))
 
 	return userId, nil
 }
